pce500/requirements.txt
numpy>=1.24.0
Pillow>=10.0.0